+++ src/test/java/com/landvibe/landlog/service/MemberServiceTest.java
@@ -1,8 +1,9 @@
 package com.landvibe.landlog.service;
 
-import com.landvibe.landlog.controller.LoginForm;
+import com.landvibe.landlog.controller.form.LoginForm;
 import com.landvibe.landlog.domain.Member;
 import com.landvibe.landlog.repository.MemoryMemberRepository;
+
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -10,8 +11,6 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import java.util.List;
 
 class MemberServiceTest {
